+++ Controller/Category.php
@@ -11,9 +11,6 @@
 
 namespace Shop\Controller;
 
-use Shop\Service\CategorySortGadget;
-use Krystal\Form\Gadget\PerPageCount;
-
 final class Category extends AbstractShopController
 {
     /**
@@ -98,47 +95,6 @@ final class Category extends AbstractShopController
         // Append breadcrumbs now
         $this->view->getBreadcrumbBag()
                    ->add($breadcrumbs);
-    }
-
-    /**
-     * Returns per page count
-     * 
-     * @return \Shop\Service\PerPageCountProvider
-     */
-    private function getPerPageCountGadget()
-    {
-        static $gadget = null;
-
-        if (is_null($gadget)) {
-            // Get default value
-            $config = $this->getModuleService('configManager')->getEntity();
-            $default = $config->getDefaultCategoryPerPageCount();
-
-            // Prepare defaults
-            $defaults = array($default, 3, 5, 10, 15, 20, 25); // Default collection
-            $defaults = array_unique($defaults); // Removed duplicates if any
-            asort($defaults, \SORT_NUMERIC); // Sort from lower to highest
-
-            $gadget = new PerPageCount($this->sessionBag, 'cat_pc', $default, $defaults);
-        }
-
-        return $gadget;
-    }
-
-    /**
-     * Returns prepared category sort provider
-     * 
-     * @return \Krystal\Service\CategorySortGadget
-     */
-    private function getCategorySortGadget()
-    {
-        static $gadget = null;
-
-        if (is_null($gadget)) {
-            $gadget = new CategorySortGadget($this->sessionBag);
-        }
-
-        return $gadget;
     }
 
     /**